requirements.txt
qwen-vl-utils
transformers
gradio
opencv-python
numpy
pillow
torch
torchvision